requirements.txt
timm>=0.6.12
# Minimal requirements for data preparation and prototyping
torch>=1.12.0
torchvision>=0.13.0
pandas>=1.3.0
scikit-learn>=1.0.2
Pillow>=8.0.0
matplotlib>=3.4.0
numpy>=1.21.0
tqdm>=4.64.0
scipy>=1.8.0

